backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
anthropic==0.18.0
pypdf2==3.0.1
python-docx==1.1.0
gspread==6.0.0
oauth2client==4.1.3
python-dotenv==1.0.0
websockets==12.0
